+++ app/controllers/admin/tags_controller.rb
@@ -1,0 +1,13 @@
+module Admin
+  class TagsController < LocalizedReferencesController
+    private
+
+    def model_class
+      Tag
+    end
+
+    def translation_association
+      :tag_translations
+    end
+  end
+end