+++ app/interactors/admin/reference_data/destroy_localized_record.rb
@@ -1,0 +1,20 @@
+module Admin
+  module ReferenceData
+    class DestroyLocalizedRecord < ApplicationInteractor
+      option :record
+
+      class ValidationContract < ApplicationContract
+        params do
+          required(:record)
+        end
+      end
+
+      def call
+        record.destroy!
+        Success(record)
+      rescue ActiveRecord::DeleteRestrictionError, ActiveRecord::InvalidForeignKey
+        Failure(code: :destroy_failed, record:)
+      end
+    end
+  end
+end